performance om nom nom nom nom

+++ TeamCode/src/main/scripts/cache/clean-old-cache/prune-to-size.js
@@ -23,7 +23,9 @@ module.exports = function(cacheMeta, cacheDir, cacheMaxBytes) {
 }
 
 function removeMetaEntry(cacheMeta, cacheMetaEntry) {
-    console.warn("Flushing cache entry " + cacheMetaEntry.key);
-    delete cacheMeta[cacheMetaEntry.key];
-    fs.unlinkSync(cacheMetaEntry.file);
+    if(cacheMetaEntry != undefined) {
+        console.warn("Flushing cache entry " + cacheMetaEntry.key);
+        delete cacheMeta[cacheMetaEntry.key];
+        fs.unlinkSync(cacheMetaEntry.file);
+    }
 }
